refactor: consolidate testing
